backend/requirements-dev.txt
# Development and Testing Dependencies for AI News Aggregator
# Comprehensive requirements for development, testing, and quality assurance

# Core Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# Coverage and Code Quality
coverage>=7.2.0
coverage-badge>=1.1.0
bandit>=1.7.5
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
flake8-import-order>=0.18.0
flake8-bugbear>=23.7.0
flake8-comprehensions>=3.14.0
flake8-simplify>=0.20.0

# Type Checking
mypy>=1.5.0
types-requests>=2.31.0
types-redis>=4.6.0
types-PyYAML>=6.0.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
mkdocs>=1.5.0
mkdocs-material>=9.1.0

# Development Tools
pre-commit>=3.3.0
commitizen>=3.8.0
semantic-release>=21.0.0
bump2version>=1.0.0

# Performance Testing
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0
psutil>=5.9.0
locust>=2.15.0

# API Testing
httpx>=0.24.0
requests>=2.31.0
pydantic>=2.1.0

# Database Testing
sqlalchemy-utils>=0.41.0
factory-boy>=3.3.0
freezegun>=1.2.0

# Mock Services
responses>=0.23.0
aioresponses>=0.7.0
aiofiles>=23.1.0

# Security Testing
safety>=2.3.0
pip-audit>=2.6.0

# Load Testing
locust>=2.15.0
artillery>=2.0.0

# Linting and Formatting
pylint>=2.17.0
pylint-django>=2.5.0
pydocstyle>=6.3.0

# Debugging
ipdb>=0.13.0
pdb++>=0.10.0
rich>=13.4.0
typer>=0.9.0

# Development Environment
python-dotenv>=1.0.0
click>=8.1.0

# Testing Utilities
faker>=19.3.0
pytest-factoryboy>=2.5.0
pytest-django>=4.5.0
pytest-celery>=0.0.0

# Monitoring and Profiling
py-spy>=0.3.0
line-profiler>=4.0.0
scalene>=1.5.0

# Documentation Generation
mkdocs-material>=9.1.0
mkdocs-mermaid2-plugin>=1.1.0
mkdocs-git-revision-date-localized-plugin>=1.2.0

# Continuous Integration
tox>=4.6.0
nox>=2023.4.0

# Quality Assurance
pylama>=8.4.0
prospector>=1.9.0
vulture>=2.7.0